+++ CHANGES.rst
@@ -3,6 +3,8 @@ ChangeLog
 
 0.3 (TBD)
 ---------
+* [COMMAND ADDED] stop command terminates docker containers
+* [COMMAND ADDED] start command initiates docker containers
 * [FEATURE ADDED] init commmand has option to setup framework for python modules
 * [FEATURE ADDED] update command updates the setup.py comments